fix(migrator-j2store4): add missing J2store4Adapter class, unblocks fatal PHP error
The inline J2store4Adapter defined inside Extension/J2store4.php had two
critical defects:
1. validatePrerequisites(?SourceDatabaseReaderInterface) signature mismatched
   the interface's validatePrerequisites() (zero args), causing a PHP fatal
   error on plugin load.
2. SourceInfo was constructed with named arg 'key' which does not exist on
   the DTO, and was missing the required 'version' arg.

Fix: extract adapter to src/Adapter/J2store4Adapter.php under its own
namespace (J2Commerce\Plugin\J2commercemigrator\J2store4\Adapter), correct
all signatures and DTO usage, port all constants verbatim from TableMapper
and DataTransformer (TABLE_MAP 51 entries, PK_REPLACEMENTS 51 entries,
JSON_FIELDS 16 entries, DATE_COLUMNS 27 entries, EXTRA_COLUMN_DEFAULTS 5
tables) per PRD §19.3–§19.6.

Extension/J2store4.php minimal change: swap unused imports for a single
use J2store4Adapter import. The registerAdapter() call site is unchanged.
